backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-genai>=1.0.0
numpy>=1.26.0
scipy>=1.12.0
Pillow>=10.2.0
trimesh>=4.0.0
manifold3d>=3.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
